refactor navigation sidebar
